add mlir binaries to bin/; factor out utils; move into packaging directory
