+++ .github/workflows/buildRyzenWheels.yml
@@ -72,7 +72,7 @@ jobs:
           export DATETIME=$(date +"%Y%m%d%H")
           export REQUIRED_LLVM_WHEEL_VERSION=$VERSION \
           
-          pushd utils/mlir_aie_wheels
+          pushd utils/packaging/mlir_aie_wheels
           
           pip install wheel auditwheel patchelf importlib_metadata
           CIBW_ARCHS=x86_64 \
@@ -173,9 +173,9 @@ jobs:
           export AIE_PROJECT_COMMIT=$(git rev-parse --short HEAD)
           export DATETIME=$(date +"%Y%m%d%H")
           
-          cp python/requirements.txt utils/mlir_aie_wheels/python_bindings
+          cp python/requirements.txt utils/packaging/mlir_aie_wheels/python_bindings
           
-          pushd utils/mlir_aie_wheels/python_bindings
+          pushd utils/packaging/mlir_aie_wheels/python_bindings
           
           pip install wheel auditwheel patchelf
           CIBW_ARCHS=x86_64 pip wheel . -v -w $WHEELHOUSE_DIR --no-build-isolation
