+++ utils/packaging/setuptools_utils.py
@@ -1,0 +1,54 @@
+import platform
+import shutil
+import os
+from pathlib import Path
+from setuptools.command.install_scripts import install_scripts, log
+
+
+# --------------------------------------------------------------------------
+# Utilities
+
+
+def check_env(build, default=0):
+    return os.getenv(build, str(default)).lower() in {"1", "true", "true", "on", "yes"}
+
+
+def get_exe_suffix():
+    if platform.system() == "Windows":
+        suffix = ".exe"
+    else:
+        suffix = ""
+    return suffix
+
+
+# --------------------------------------------------------------------------
+# InstallBin command -- overwrites "install_scripts" stage of "install" cmd
+# Copy files into the <packagename>.data/data/scripts directory.
+# The wheel package format prescribes that files in <packagename>.data/scripts be added to the PATH on install.
+# So whatever is added here will end up in the virtualenv's bin directory upon install.
+class InstallBin(install_scripts):
+
+    # The subdirectory of the wheel directory in which binaries can be found
+    bin_dir = "bin"
+
+    def run(self):
+        # Copy the built binaries back into the source directory.
+        # This is needed so that the setup(data_files = [(scripts, ...)]) argument below can
+        # pick them up.
+
+        install_dir = Path(self.install_dir)
+
+        # Not a super clean way of doing this, but:
+        #  1. `build_ext`` will call CMake to build mlir_aie in <temp_build_dir>/lib.linux-x86_64.../mlir_aie
+        #  2. then, `install_lib`` will copy everything in there to <wheelhouse>/mlir_aie
+        #  3. then, this runs -- `install_scripts`, and it will be configured with self.install_dir == <wheelhouse>/<pkgname>.data/scripts
+        #     therefore, we can get our hands on the binaries moved into the wheelhouse root by looking at the grandparent dir
+        bin_dir = install_dir.parent.parent.resolve() / bin_dir
+
+        if not install_dir.exists():
+            install_dir.mkdir(parents=True)
+
+        for bin in bin_dir.glob("*"):
+            dest = str(Path(self.install_dir) / bin.name)
+            log.info(f"Copying binary {bin} -> {dest}")
+            shutil.copyfile(bin, dest)
